pyproject.toml
[project]
name = "ml_dl_experiments"
version = "0.1.0"
description = "This is my ml_dl libraries and hardcode for AI"
authors = [
    {name = "Ollldman",email = "dodrolla@gmail.com"}
]
packages = [
    { include = "ml_dl_experiments" },
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "annotated-types (==0.7.0)",
    "colorama (==0.4.6)",
    "contourpy (==1.3.3)",
    "cycler (==0.12.1)",
    "filelock (==3.19.1)",
    "fonttools (==4.60.0)",
    "fsspec (==2025.9.0)",
    "iniconfig (==2.1.0)",
    "jinja2 (==3.1.6)",
    "joblib (==1.5.2)",
    "kiwisolver (==1.4.9)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.10.6)",
    "mpmath (==1.3.0)",
    "networkx (==3.5)",
    "numpy (==2.2.6)",
    "packaging (==25.0)",
    "pandas (==2.3.2)",
    "pillow (==11.3.0)",
    "pluggy (==1.6.0)",
    "pydantic (==2.11.9)",
    "pydantic-settings (==2.11.0)",
    "pydantic-core (==2.33.2)",
    "pygments (==2.19.2)",
    "pyparsing (==3.2.5)",
    "pytest (==8.4.2)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.1)",
    "pytz (==2025.2)",
    "scikit-learn (==1.7.2)",
    "scipy (==1.16.2)",
    "seaborn (==0.13.2)",
    "setuptools (==80.9.0)",
    "six (==1.17.0)",
    "sympy (==1.14.0)",
    "threadpoolctl (==3.6.0)",
    "torch (==2.9.0)",
    "torchvision (==0.24.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.2)",
    "xgboost (==3.0.5)",
    "ipykernel (>=7.0.1,<8.0.0)",
    "ipython (>=9.6.0,<10.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "certifi (==2025.4.26)",
    "charset-normalizer (==3.4.2)",
    "gensim (>=4.4.0,<5.0.0)",
    "hf-xet (>=1.1.10,<2.0.0)",
    "huggingface-hub (>=0.36.0,<0.37.0)",
    "idna (==3.10)",
    "pyyaml (==6.0.2)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "safetensors (==0.5.3)",
    "smart-open (==7.1.0)",
    "tokenizers (>=0.22.1,<0.23.0)",
    "tqdm (==4.67.1)",
    "urllib3 (==2.4.0)",
    "wrapt (==1.17.2)",
    "clearml (>=2.0.2,<3.0.0)",
    "nvidia-cublas (==13.0.0.19)",
    "nvidia-cuda-cupti (==13.0.48)",
    "nvidia-cuda-nvrtc (==13.0.48)",
    "nvidia-cuda-runtime (==13.0.48)",
    "nvidia-cudnn-cu13 (==9.13.0.50)",
    "nvidia-cufft (==12.0.0.15)",
    "nvidia-cufile (==1.15.0.42)",
    "nvidia-curand (==10.4.0.35)",
    "nvidia-cusolver (==12.0.3.29)",
    "nvidia-cusparse (==12.6.2.49)",
    "nvidia-cusparselt-cu13 (==0.8.0)",
    "nvidia-nccl-cu13 (==2.27.7)",
    "nvidia-nvjitlink (==13.0.39)",
    "nvidia-nvshmem-cu13 (==3.3.24)",
    "nvidia-nvtx (==13.0.39)",
    "torchsummary (>=1.5.1,<2.0.0)",
    "datasets (>=4.3.0,<5.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "mpl-axes-aligner (>=1.3,<2.0)",
    "kagglehub (>=0.3.13,<0.4.0)",
    "timm (>=1.0.22,<2.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "pymorphy3 (>=2.0.6,<3.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "albumentations (>=2.0.8,<3.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "torchmetrics[detection] (>=1.8.2,<2.0.0)",
    "ultralytics (>=8.3.241,<9.0.0)",
    "pandoc (>=2.4,<3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
